+++ docs/CHANGES.md
@@ -2,6 +2,22 @@
 
 ------
 
+## [build20210712]
+
+### Features
+
+* new `eventScript` parameter
+
+
+
+## [build2021011511]
+
+### Features
+
+* Add parameter `[notify]onlyChanges`
+
+
+
 ## [build202101162]
 
 ### Fixed